fix final issues and duplication
