+++ api/src/queries/project/project-view-queries.ts
@@ -367,37 +367,6 @@ export const getFundingSourceByProjectSQL = (projectId: number): SQLStatement | 
 };
 
 /**
- * SQL query to get project stakeholder partnerships.
- * @param {number} projectId
- * @returns {SQLStatement} sql query object
- */
-export const getStakeholderPartnershipsByProjectSQL = (projectId: number): SQLStatement | null => {
-  defaultLog.debug({ label: 'getStakeholderPartnershipsByProjectSQL', message: 'params', projectId });
-
-  if (!projectId) {
-    return null;
-  }
-
-  const sqlStatement = SQL`
-    SELECT
-      name
-    FROM
-      stakeholder_partnership
-    WHERE
-      p_id = ${projectId};
-  `;
-
-  defaultLog.debug({
-    label: 'getStakeholderPartnershipsByProjectSQL',
-    message: 'sql',
-    'sqlStatement.text': sqlStatement.text,
-    'sqlStatement.values': sqlStatement.values
-  });
-
-  return sqlStatement;
-};
-
-/**
  * SQL query to get project indigenous partnerships.
  * @param {number} projectId
  * @returns {SQLStatement} sql query object
